chore(xlsx): extract clone-side axis resolvers (resolveAxes + applyOverride helpers) into chart/axis
Move resolveAxes (the 772-line clone-side axis orchestrator) and every
apply{Skip,LblOffset,LblAlgn,NoMultiLvlLbl,Auto,Hidden,Gridlines,Scale,
NumberFormat,TickMark,TickLblPos,LabelRotation,LabelFontSize,LabelBold,
LabelItalic,LabelColor,LabelUnderline,LabelStrike,LabelFontFamily,
AxisTitle*,Reverse,Crosses,DispUnits,CrossBetween}Override helper from
chart-clone.ts into chart/axis.ts. Move the cloneScale,
clampLabelRotationDeg, normalizeLabelFontFamily,
normalizeAxisTitleFontFamilyClone, resolveAxisTitleLayout, and
normalizeDispUnits helpers along with them, plus the clone-side
VALID_TICK_MARK_VALUES / VALID_TICK_LBL_POS_VALUES /
VALID_CROSSES_VALUES / VALID_DISP_UNIT_VALUES /
VALID_CROSS_BETWEEN_VALUES enums and the CrossesPair{Source,Override}
interfaces.

Rename the clone-side resolveAutoTitleDeleted to
resolveCloneAutoTitleDeleted so it does not collide with the
writer-side single-arg resolveAutoTitleDeleted. The new name is
private to the chart/clone code path and not part of the public API.

Move the clone-side resolveBorderWidthPt and resolveBorderDash helpers
into chart/shape.ts since their accept/normalize grammar is shape-
generic; they are now reused by chart/axis.ts as well as chart-clone.ts.

Lift the clone-side applyOverride generic into chart/util.ts (typed as
`applyOverride<T>(source, override): T | undefined`) so per-host modules
can compose source/override resolvers without re-implementing the
inherit/drop/replace dispatch.

Pure relocation — JSDoc preserved verbatim, no behavior change. All
7362 tests still pass; typecheck clean.
